+++ src/Resources/contao/languages/de/tl_list_config_element.php
@@ -5,6 +5,4 @@ $lang = &$GLOBALS['TL_LANG']['tl_list_config_element'];
 /**
  * Reference
  */
-$lang['reference'] += [
-    \HeimrichHannot\GoogleMapBundle\ConfigElementType\GoogleMapListConfigElementType::TYPE => 'Google Map',
-];
+$lang['reference'][\HeimrichHannot\GoogleMapBundle\ConfigElementType\GoogleMapConfigElementType::TYPE] = 'Google Map';